# Conflicts:
#	src/main/java/africa/nkwadoma/nkwadoma/application/ports/output/identity/IdentityManagerOutPutPort.java

+++ src/main/java/africa/nkwadoma/nkwadoma/infrastructure/adapters/output/persistence/education/ProgramPersistenceAdapter.java
@@ -10,14 +10,16 @@ import africa.nkwadoma.nkwadoma.domain.model.education.*;
 import africa.nkwadoma.nkwadoma.domain.model.identity.*;
 import africa.nkwadoma.nkwadoma.domain.validation.*;
 import africa.nkwadoma.nkwadoma.infrastructure.adapters.output.mapper.ProgramMapper;
-import africa.nkwadoma.nkwadoma.infrastructure.adapters.output.persistence.entity.education.ProgramEntity;
+import africa.nkwadoma.nkwadoma.infrastructure.adapters.output.persistence.entity.education.*;
 import africa.nkwadoma.nkwadoma.infrastructure.adapters.output.persistence.entity.organization.*;
 import africa.nkwadoma.nkwadoma.infrastructure.adapters.output.persistence.mapper.OrganizationIdentityMapper;
 import africa.nkwadoma.nkwadoma.infrastructure.adapters.output.persistence.repository.*;
 import africa.nkwadoma.nkwadoma.infrastructure.adapters.output.persistence.repository.education.*;
+import africa.nkwadoma.nkwadoma.infrastructure.adapters.output.persistence.repository.education.ProgramRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections.*;
+import org.apache.commons.lang3.*;
 import org.springframework.data.domain.*;
 import org.springframework.stereotype.Component;
 
@@ -26,6 +28,7 @@ import java.util.*;
 
 import static africa.nkwadoma.nkwadoma.domain.enums.constants.ProgramMessages.PROGRAM_NOT_FOUND;
 import static africa.nkwadoma.nkwadoma.domain.validation.MeedlValidator.validateDataElement;
+import static africa.nkwadoma.nkwadoma.domain.validation.MeedlValidator.validateUUID;
 
 @RequiredArgsConstructor
 @Component
@@ -37,11 +40,13 @@ public class ProgramPersistenceAdapter implements ProgramOutputPort {
     private final OrganizationIdentityMapper organizationIdentityMapper;
     private final OrganizationEntityRepository organizationEntityRepository;
     private final OrganizationEmployeeIdentityOutputPort employeeIdentityOutputPort;
+    private final CohortRepository cohortRepository;
 
     @Override
     public Program findProgramByName(String programName) throws MeedlException {
         validateDataElement(programName);
         programName = programName.trim();
+        log.info("Program being searched for by name is {}", programName);
         ProgramEntity programEntity = programRepository.findByName(programName).
                 orElseThrow(()-> new ResourceNotFoundException(PROGRAM_NOT_FOUND.getMessage()));
         return programMapper.toProgram(programEntity);
@@ -53,7 +58,9 @@ public class ProgramPersistenceAdapter implements ProgramOutputPort {
         program.validate();
         validateCreatedBy(program);
 
+        log.info("Program at persistence layer: ========>{}", program);
         OrganizationIdentity organizationIdentity = organizationIdentityOutputPort.findById(program.getOrganizationId());
+        log.info("The organization identity found when saving program is: {}", organizationIdentity);
         List<ServiceOffering> serviceOfferings = organizationIdentityOutputPort.findServiceOfferingById(organizationIdentity.getId());
         ProgramPersistenceAdapter.validateServiceOfferings(serviceOfferings);
 
@@ -62,17 +69,21 @@ public class ProgramPersistenceAdapter implements ProgramOutputPort {
         ProgramEntity programEntity = programMapper.toProgramEntity(program);
         programEntity.setOrganizationEntity(organizationEntity);
         programEntity = programRepository.save(programEntity);
+        updateOrganization(program, organizationEntity);
 
-        organizationEntity.setNumberOfPrograms(organizationEntity.getNumberOfPrograms() + BigInteger.ONE.intValue());
-        organizationEntityRepository.save(organizationEntity);
         return programMapper.toProgram(programEntity);
     }
 
-    private void validateCreatedBy(Program program) throws MeedlException {
-        Optional<OrganizationEmployeeIdentity> employeeIdentity = employeeIdentityOutputPort.findByCreatedBy(program.getCreatedBy());
-        if (employeeIdentity.isEmpty()) {
-            throw new IdentityException(MeedlMessages.NON_EXISTING_CREATED_BY.getMessage());
+    private void updateOrganization(Program program, OrganizationEntity organizationEntity) {
+        if (StringUtils.isEmpty(program.getId())) {
+            organizationEntity.setNumberOfPrograms(organizationEntity.getNumberOfPrograms() + BigInteger.ONE.intValue());
+            organizationEntityRepository.save(organizationEntity);
         }
+    }
+
+    private void validateCreatedBy(Program program) throws MeedlException {
+        log.info("Validating the created by: {}",program.getCreatedBy());
+        OrganizationEmployeeIdentity employeeIdentity = employeeIdentityOutputPort.findByCreatedBy(program.getCreatedBy());
     }
 
 
@@ -84,15 +95,22 @@ public class ProgramPersistenceAdapter implements ProgramOutputPort {
 
     @Override
     public void deleteProgram(String programId) throws MeedlException {
-        validateDataElement(programId);
+        MeedlValidator.validateDataElement(programId);
+        MeedlValidator.validateUUID(programId);
+        Optional<CohortEntity> cohortEntity = cohortRepository.findByProgramId(programId);
+        if (cohortEntity.isPresent()) {
+            throw new EducationException(ProgramMessages.COHORT_EXISTS.getMessage());
+        }
         ProgramEntity program = programRepository.findById(programId).
-                orElseThrow(() -> new ResourceNotFoundException(PROGRAM_NOT_FOUND.getMessage()));
+                orElseThrow(() -> new ResourceNotFoundException(ProgramMessages.PROGRAM_NOT_FOUND.getMessage()));
         programRepository.delete(program);
     }
 
     @Override
     public Program findProgramById(String programId) throws MeedlException {
-        validateDataElement(programId);
+        MeedlValidator.validateDataElement(programId);
+        MeedlValidator.validateUUID(programId);
+        programId = programId.trim();
         ProgramEntity programEntity = programRepository.findById(programId).
                 orElseThrow(() -> new ResourceNotFoundException(PROGRAM_NOT_FOUND.getMessage()));
         return programMapper.toProgram(programEntity);
@@ -104,8 +122,8 @@ public class ProgramPersistenceAdapter implements ProgramOutputPort {
         Page<ProgramEntity> programEntities = programRepository.findAllByOrganizationEntityId(organizationId, pageRequest);
         return programEntities.map(programMapper::toProgram);
     }
-
     private static void validateServiceOfferings(List<ServiceOffering> serviceOfferings) throws EducationException {
+        log.info("Validating service offerings: {}", serviceOfferings);
         if(CollectionUtils.isEmpty(serviceOfferings) ||
                 !serviceOfferings.stream().map(ServiceOffering::getName).toList().contains(ServiceOfferingType.TRAINING.name())) {
             throw new EducationException(ProgramMessages.INVALID_SERVICE_OFFERING.getMessage());
